Updating UI (#79)
* change medley

* format

* Find on input dialog

* expand menu options

* func fix

* remove comment

* format

* y-axis titles

* requested changes and support for button menu

* format

* menu fix

* merge fixes

* build fix

format

Find on input dialog

expand menu options

func fix

remove comment

format

y-axis titles

requested changes and support for button menu

format

menu fix

Commenting organisation and grouping fixes

rename outdated funcs + remove obsolete elems

fix action shortcuts

build reset

merge fixes
