+++ frontend/mainwindow.h
@@ -26,50 +26,56 @@ class MainWindow : public QMainWindow
     public:
     MainWindow(QWidget *parent = nullptr);
     ~MainWindow();
-
+    // Init
     void fillInstruments(QList<QPair<QString, QString>> instruments);
     void initialiseElements();
-    void goToCurrentFoundIndex(QModelIndex index);
-    QList<QPair<QString, QString>> getInstruments();
-    std::vector<std::pair<QString, QString>> getFields(QString instrument, QString instType);
+    // Misc
+    void goToCurrentFoundIndex(QModelIndex index); // Selects given index
+    QList<QPair<QString, QString>> getInstruments(); // Get Instruments from config file
+    std::vector<std::pair<QString, QString>> getFields(QString instrument, QString instType); // Get Fields from config file
     void setLoadScreen(bool state);
     private slots:
+    // Search Controls
+    void updateSearch(const QString &arg1);
+    void on_actionSearch_triggered();
+    void on_actionSelectNext_triggered();
+    void on_actionSelectPrevious_triggered();
+    void on_actionSelectAll_triggered();
+    void findUp();
+    void findDown();
+    void selectAllSearches();
+    void selectIndex(QString runNumber);
+    void selectSimilar();
+    // Filter Controls
     void on_filterBox_textChanged(const QString &arg1);
-    void on_searchBox_textChanged(const QString &arg1);
-    void handle_result_instruments(HttpRequestWorker *worker);
-    void handle_result_cycles(HttpRequestWorker *worker);
-    void on_instrumentsBox_currentTextChanged(const QString &arg1);
-    void on_cyclesBox_currentTextChanged(const QString &arg1);
-    void on_groupButton_clicked(bool checked);
-    void columnHider(int state);
     void on_clearSearchButton_clicked();
-    void on_findUp_clicked();
-    void on_findDown_clicked();
-    void on_searchAll_clicked();
-    void on_closeFind_clicked();
-    void recentCycle();
-    void customMenuRequested(QPoint pos);
-    void handle_result_contextGraph(HttpRequestWorker *worker);
-    void contextGraph();
-    void handle_result_contextMenu(HttpRequestWorker *worker);
-    void removeTab(int index);
-    void toggleAxis(int state);
-    void savePref();
-    void showStatus(qreal x, qreal y);
     void massSearch(QString name, QString value);
     void on_actionMassSearchRB_No_triggered();
     void on_actionMassSearchTitle_triggered();
     void on_actionMassSearchUser_triggered();
     void on_actionClear_cached_searches_triggered();
     void goTo(HttpRequestWorker *worker, QString runNumber);
-    void selectIndex(QString runNumber);
-    void selectSimilar();
-    void changeInst(QPair<QString, QString> instrument);
-    void on_actionSearch_triggered();
-    void on_actionSelectNext_triggered();
-    void on_actionSelectPrevious_triggered();
-    void on_actionSelectAll_triggered();
     void on_actionRun_Number_triggered();
+    // Data Selection
+    void handle_result_instruments(HttpRequestWorker *worker);
+    void handle_result_cycles(HttpRequestWorker *worker);
+    void currentInstrumentChanged(const QString &arg1);
+    void on_cyclesBox_currentTextChanged(const QString &arg1);
+    void recentCycle();
+    void changeInst(QPair<QString, QString> instrument);
+    // Grouping
+    void on_groupButton_clicked(bool checked);
+    // Visualisation
+    void customMenuRequested(QPoint pos);
+    void handle_result_contextGraph(HttpRequestWorker *worker);
+    void contextGraph();
+    void handle_result_contextMenu(HttpRequestWorker *worker);
+    void toggleAxis(int state);
+    void showStatus(qreal x, qreal y);
+    // Misc Interface Functions
+    void removeTab(int index);
+    void savePref();
+    void columnHider(int state);
 
     protected:
     // Window close event
@@ -81,18 +87,25 @@ class MainWindow : public QMainWindow
 
     private:
     Ui::MainWindow *ui_;
+    // Table Stuff
     JsonTableModel *model_;
     QSortFilterProxyModel *proxyModel_;
+    JsonTableModel::Header header_;
+    std::vector<std::pair<QString, QString>> desiredHeader_;
+    // Menus
     QMenu *viewMenu_;
     QMenu *findMenu_;
     QMenu *contextMenu_;
     QMenu *instrumentsMenu_;
-    JsonTableModel::Header header_;
-    std::vector<std::pair<QString, QString>> desiredHeader_;
+    
     QModelIndexList foundIndices_;
     int currentFoundIndex_;
+    // Menu button data
+    QString searchString_;
+    QString instType_;
+    QString instName_;
+    // Misc
     bool init_;
-    QChart *chart_;
     QPoint pos_;
     QList<std::tuple<HttpRequestWorker *, QString>> cachedMassSearch_;
     QString searchString_;
