+++ frontend/searching.cpp
@@ -4,9 +4,10 @@
 #include "./ui_mainwindow.h"
 #include "mainwindow.h"
 #include <tuple>
+#include <QInputDialog>
 
 // Search table data
-void MainWindow::on_searchBox_textChanged(const QString &arg1)
+void MainWindow::updateSearch(const QString &arg1)
 {
     foundIndices_.clear();
     currentFoundIndex_ = 0;
@@ -38,7 +39,7 @@ void MainWindow::on_searchBox_textChanged(const QString &arg1)
 }
 
 // Select previous match
-void MainWindow::on_findUp_clicked()
+void MainWindow::findUp()
 {
     // Boundary/ error handling
     if (foundIndices_.size() > 0)
@@ -54,7 +55,7 @@ void MainWindow::on_findUp_clicked()
 }
 
 // Select next match
-void MainWindow::on_findDown_clicked()
+void MainWindow::findDown()
 {
     // Boundary/ error handling
     if (foundIndices_.size() > 0)
@@ -68,7 +69,7 @@ void MainWindow::on_findDown_clicked()
 }
 
 // Select all matches
-void MainWindow::on_searchAll_clicked()
+void MainWindow::selectAllSearches()
 {
     // Error handling
     if (foundIndices_.size() > 0)
@@ -89,3 +90,68 @@ void MainWindow::goToCurrentFoundIndex(QModelIndex index)
     ui_->runDataTable->selectionModel()->setCurrentIndex(index,
                                                          QItemSelectionModel::ClearAndSelect | QItemSelectionModel::Rows);
 }
+
+void MainWindow::selectIndex(QString runNumber)
+{
+    ui_->runDataTable->selectionModel()->clearSelection();
+
+    updateSearch(runNumber);
+    statusBar()->showMessage("Found run " + runNumber + " in " + ui_->cyclesBox->currentText(), 5000);
+    disconnect(this, &MainWindow::tableFilled, nullptr, nullptr);
+}
+
+void MainWindow::selectSimilar()
+{
+    int TitleColumn;
+    for (auto i = 0; i < ui_->runDataTable->horizontalHeader()->count(); ++i)
+    {
+        if (model_->headerData(i, Qt::Horizontal, Qt::UserRole).toString() == "title")
+        {
+            TitleColumn = i;
+            break;
+        }
+    }
+    QString title = model_->index(ui_->runDataTable->rowAt(pos_.y()), TitleColumn).data().toString();
+    for (auto i = 0; i < model_->rowCount(); i++)
+    {
+        if (model_->index(i, TitleColumn).data().toString() == title)
+            ui_->runDataTable->selectionModel()->setCurrentIndex(model_->index(i, TitleColumn),
+                                                                 QItemSelectionModel::Select | QItemSelectionModel::Rows);
+    }
+}
+void MainWindow::on_actionSearch_triggered()
+{
+    QString textInput = QInputDialog::getText(this, tr("Enter search query"), tr("search runs for:"), QLineEdit::Normal);
+    searchString_ = textInput;
+    foundIndices_.clear();
+    currentFoundIndex_ = 0;
+    if (textInput.isEmpty())
+    {
+        ui_->runDataTable->selectionModel()->clearSelection();
+        statusBar()->clearMessage();
+        return;
+    }
+    // Find all occurences of search string in table elements
+    for (auto i = 0; i < proxyModel_->columnCount(); ++i)
+    {
+        auto location = ui_->runDataTable->horizontalHeader()->logicalIndex(i);
+        if (ui_->runDataTable->isColumnHidden(location) == false)
+            foundIndices_.append(
+                proxyModel_->match(proxyModel_->index(0, location), Qt::DisplayRole, textInput, -1, Qt::MatchContains));
+    }
+    // Select first match
+    if (foundIndices_.size() > 0)
+    {
+        goToCurrentFoundIndex(foundIndices_[0]);
+        statusBar()->showMessage("Find \"" + searchString_ + "\": 1/" + QString::number(foundIndices_.size()) + " Results");
+    }
+    else
+    {
+        ui_->runDataTable->selectionModel()->clearSelection();
+        statusBar()->showMessage("No results");
+    }
+}
+
+void MainWindow::on_actionSelectNext_triggered() { findDown(); }
+void MainWindow::on_actionSelectPrevious_triggered() { findUp(); }
+void MainWindow::on_actionSelectAll_triggered() { selectAllSearches(); }
